add mock browser to test session checking logic
