+++ auth/mock.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"time"
 
-	"github.com/stretchr/testify/mock"
 	"golang.org/x/oauth2"
 )
 
@@ -15,11 +14,19 @@ func (s *MockSession) GetToken(ctx context.Context, conf *oauth2.Config) (*oauth
 	tok := oauth2.Token{
 		AccessToken:  "SomeAccessToken",
 		RefreshToken: "SomeRefreshToken",
-		Expiry:       time.Now(),
+		Expiry:       time.Now().Add(time.Hour * 1),
 	}
 	return &tok, nil
 }
 
-type MockBrowser struct {
-	mock.Mock
+// mockBrowser provides a mocked response of the OAuth2 login flow.
+type mockBrowser struct{}
+
+func (b *mockBrowser) GetTokenFromBrowser(ctx context.Context, conf *oauth2.Config) (*oauth2.Token, error) {
+	tok := oauth2.Token{
+		AccessToken:  "SomeNewAccessToken",
+		RefreshToken: "SomeNewRefreshToken",
+		Expiry:       time.Now().Add(time.Hour * 1),
+	}
+	return &tok, nil
 }
